requirements.txt
beautifulsoup4==4.11.1
bs4==0.0.1
charset-normalizer==2.1.1
click
colorama
dnspython==2.2.1
Flask
Flask-Cors==3.0.10
idna==3.4
itsdangerous
Jinja2 
MarkupSafe
mysql-connector-python==8.0.31
protobuf==3.20.1
pymongo==4.3.3
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
urllib3==1.26.13
Werkzeug
wincertstore==0.2
python-decouple
